requirements.txt
Django>=6.0,<7.0
graphene-django>=3.2.0
django-filter>=25.0

